backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pypdf==5.1.0
langchain==0.3.7
langchain-core==0.3.15
langchain-groq==0.2.3
langchain-chroma==0.1.4
langchain-huggingface==0.1.0
langchain-text-splitters==0.3.2
langgraph==0.2.45
chromadb==0.5.18
sentence-transformers==3.2.1
pydantic==2.9.2
python-dotenv==1.0.1
reportlab==4.2.5
